+++ terraform/routes.tf
@@ -21,3 +21,17 @@ resource "cloudfoundry_route" "beis-rpr-redirect-route" {
   space    = cloudfoundry_space.space.id
   hostname = var.custom_hostname
 }
+
+# Create a route for the app using the redirect domain
+# not using hostname to allow the redirection of the root domain - only for production
+resource "cloudfoundry_route" "beis-rpr-redirect-root-route" {
+  count       = var.environment == "prod" ? 1 : 0
+  domain      = data.cloudfoundry_domain.redirect.id
+  space       = cloudfoundry_space.space.id
+  depends_on  = [
+    cloudfoundry_app.beis-rpr-redirect 
+  ]
+  target {
+    app   = cloudfoundry_app.beis-rpr-redirect.id
+  }
+}